+++ lib/src/annotations.dart
@@ -5,3 +5,9 @@ class Live {
 class LiveStore {
   const LiveStore();
 }
+
+class Computed {
+  const Computed();
+}
+
+const computed = Computed();